Move class TInflightInfo to separate file
