+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/dirty_map/ut/ya.make
@@ -1,8 +1,8 @@
 UNITTEST_FOR(ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/dirty_map)
 
 SRCS(
-    dirty_map_inflight_ut.cpp
     dirty_map_ut.cpp
+    inflight_info_ut.cpp
     location_ut.cpp
     range_locker_ut.cpp
 )
